+++ src/main/java/com/school/dinosaur_api/domain/service/StudentService.java
@@ -1,6 +1,7 @@
 package com.school.dinosaur_api.domain.service;
 
 import com.school.dinosaur_api.domain.exception.BusinessException;
+import com.school.dinosaur_api.domain.exception.ResourceNotFoundException;
 import com.school.dinosaur_api.domain.model.Student;
 import com.school.dinosaur_api.domain.repository.StudentRepository;
 import lombok.AllArgsConstructor;
@@ -16,17 +17,38 @@ public class StudentService {
 
     public Student findStudent(Long studentId) {
         return studentRepository.findById(studentId)
-                .orElseThrow(() -> new BusinessException("Student not found"));
+                .orElseThrow(() -> new ResourceNotFoundException("Student not found with id: " + studentId));
     }
 
     @Transactional
-    public Student saveStudent(Student student) {
+    public Student createStudent(Student student) {
+        if (student.getId() != null) {
+            throw new BusinessException("Must not contain the ID in the request body");
+        }
+
         boolean cpfUsed = studentRepository.findByCpf(student.getCpf())
                 .filter(s -> !s.equals(student))
                 .isPresent();
 
         if (cpfUsed) {
-            throw new BusinessException("CPF already in use");
+            throw new BusinessException("CPF " + student.getCpf() + " already in use");
+        }
+
+        return studentRepository.save(student);
+    }
+
+    @Transactional
+    public Student updateStudent(Student student) {
+        if (!studentRepository.existsById(student.getId())) {
+            throw new ResourceNotFoundException("Student not found with id: " + student.getId());
+        }
+
+        boolean cpfUsed = studentRepository.findByCpf(student.getCpf())
+                .filter(s -> !s.equals(student))
+                .isPresent();
+
+        if (cpfUsed) {
+            throw new BusinessException("CPF " + student.getCpf() + " already in use");
         }
 
         return studentRepository.save(student);
@@ -34,6 +56,10 @@ public class StudentService {
 
     @Transactional
     public void deleteStudent(Long studentId) {
+        if (!studentRepository.existsById(studentId)) {
+            throw new ResourceNotFoundException("Student not found with id: " + studentId);
+        }
+
         studentRepository.deleteById(studentId);
     }
 }
